styles: Change styles of login page and toast component
