+++ src/components/Toast/index.jsx
@@ -3,7 +3,7 @@ import * as ToastComponent from '@radix-ui/react-toast'
 
 import styles from './styles.module.css'
 
-export function Toast({ titulo, descricao, isError = false }) {
+export function Toast({ title, description, isError = false }) {
   const [open, setOpen] = useState(true)
 
   return (
@@ -14,10 +14,10 @@ export function Toast({ titulo, descricao, isError = false }) {
         onOpenChange={setOpen}
       >
         <ToastComponent.Title className={styles.title}>
-          {titulo}
+          {title}
         </ToastComponent.Title>
         <ToastComponent.Description asChild>
-          <p className={styles.description}>{descricao}</p>
+          <p className={styles.description}>{description}</p>
         </ToastComponent.Description>
       </ToastComponent.Root>
       <ToastComponent.Viewport className={styles.viewPort} />
