+++ src/components/Toast/styles.module.css
@@ -1,7 +1,8 @@
 .toast {
-  background-color: #28a745;
+  background-color: var(--gray-10);
+  border: 2px solid var(--green-3);
   border-radius: 8px;
-  box-shadow: hsl(206 22% 7% / 35%) 0px 10px 38px -10px, hsl(206 22% 7% / 20%) 0px 10px 20px -15px;
+  box-shadow: hsla(210, 60%, 64%, 0.35) 0px 10px 38px -10px, hsl(206 22% 7% / 20%) 0px 10px 20px -15px;
   padding: 1rem;
   display: grid;
   column-gap: 1rem;
@@ -9,7 +10,8 @@
 }
 
 .toastError {
-  background-color: var(--red-400);
+  background-color: var(--gray-10);
+  border: 2px solid var(--red);
   border-radius: 8px;
   box-shadow: hsl(206 22% 7% / 35%) 0px 10px 38px -10px, hsl(206 22% 7% / 20%) 0px 10px 20px -15px;
   padding: 1rem;
