+++ src/global.css
@@ -46,8 +46,8 @@
 }
 
 body {
-  background: var(--blue-500);
-  color: var(--white);
+  background: var(--gray-11);
+  color: var(--gray-1);
   --webkit-font-smoothing: antialiased;
 }
