--- src/pages/public/Login/styles.module.css
@@ -1,35 +1,0 @@
-.container {
-  max-width: 75rem;
-  width: 100%;
-  padding: 3.75rem 1.25rem;
-  margin: auto;
-}
-.container h1 {
-  font-size: 2.5rem;
-  line-height: 1.1;
-  font-weight: 800;
-  margin-bottom: 2.5rem;
-  text-align: center;
-}
-
-.form {
-  background-color: rgba(30, 30, 30, 0.2);
-  padding: 2.5rem;
-  border-radius: 5px;
-  display: flex;
-  flex-direction: column;
-  gap: 1.25rem;
-  max-width: 35rem;
-  margin: 0 auto;
-}
-
-.passwordRecovery {
-  color: var(--white);
-  text-decoration: none;
-  align-self: flex-end;
-  transition: 0.3s;
-  margin-bottom: 1.25rem;
-}
-.passwordRecovery:hover {
-  color: var(--blue-300);
-}
