+++ src/pages/public/SignIn/styles.module.css
@@ -1,0 +1,86 @@
+.container {
+  max-width: calc(100vw - (100vw - 1200px) / 2);
+  width: 100%;
+  padding: 0 1.25rem 0 0;
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 2.5rem;
+}
+
+.container img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.form {
+  padding: 5rem 0;
+}
+
+.form > h1 {
+  margin-bottom: 2.5rem;
+
+  color: var(--white);
+  font-weight: 800;
+  font-size: 3rem;
+  line-height: 1.1;
+}
+
+.form > form {
+  display: grid;
+  gap: 1.25rem;
+}
+
+.form > form > button {
+  justify-self: start;
+  margin-top: 1.25rem;
+}
+
+.form > a {
+  display: block;
+  margin-top: 1.25rem;
+  width: fit-content;
+  justify-self: start;
+
+  color: var(--gray-1);
+  font-weight: 500;
+  text-decoration: underline;
+  transition: 0.3s;
+}
+.form > a:hover{
+  color: var(--green-3);
+}
+
+
+
+@media(max-width: 960px) {
+  .container {
+    grid-template-columns: 1fr 2fr;
+    gap: 1.25rem;
+  }
+}
+
+@media (max-width: 600px) {
+  .container {
+    padding: 0;
+    grid-template-columns: 1fr;
+    align-content: start;
+  }
+
+  .container > div:first-child {
+    width: 100%;
+    height: 15rem;
+  }
+
+  .form {
+    padding: 0 1.25rem 2.5rem;
+  }
+
+  .form > h1 {
+    margin-bottom: 1.25rem;
+  }
+
+  .form > form > button {
+    margin-top: 0;
+  }
+}
